chore: remove deprecated ontology processing code
