+++ .github/workflows/ontology_process.yml
@@ -1,10 +1,9 @@
-name: Updates to Ontology Files
+name: Updates to GENCODE Files
 
 on:
   push:
     paths:
-      - '**/cellxgene_schema_cli/cellxgene_schema/ontology_files/gene_info.yml'
-      - '**/cellxgene_schema_cli/cellxgene_schema/ontology_files/owl_info.yml'
+      - '**/cellxgene_schema_cli/cellxgene_schema/gencode_files/gene_info.yml'
     branches-ignore:
       - main
 
@@ -16,15 +15,6 @@ jobs:
         uses: actions/checkout@v3
         with:
           ref: ${{ github.event.pull_request.head.ref }}
-      - name: ontology changes
-        uses: dorny/paths-filter@v2
-        id: filter
-        with:
-          filters: |
-            owl_info:
-              - 'cellxgene_schema_cli/cellxgene_schema/ontology_files/owl_info.yml'
-            gene_info:
-              - 'cellxgene_schema_cli/cellxgene_schema/ontology_files/gene_info.yml'
       - name: Set up Python 3.8
         uses: actions/setup-python@v1
         with:
@@ -43,21 +33,14 @@ jobs:
         run: |
           git config user.name github-actions
           git config user.email github-actions@github.com
-      - name: owl-processing
-        if: ${{ steps.filter.outputs.owl_info == 'true' }}
-        run: |
-          make download-ontologies -C cellxgene_schema_cli
-          git add ./cellxgene_schema_cli/cellxgene_schema/ontology_files/all_ontology.json.gz
       - name: gene-processing
-        if: ${{ steps.filter.outputs.gene_info == 'true' }}
         run: |
           make gene-processing -C cellxgene_schema_cli
           git add ./cellxgene_schema_cli/cellxgene_schema/ontology_files/genes_*.csv.gz
           git add ./cellxgene_schema_cli/cellxgene_schema/ontology_files/*_diff.txt
       - name: Commit
-        if: ${{ steps.filter.outputs.gene_info == 'true'  || steps.filter.outputs.owl_info == 'true' }}
         run: |
-          git commit -m "AUTO: update ontologies"
+          git commit -m "AUTO: update gencode files"
           git push
 
   unit-tests:
