+++ .github/workflows/schema-version-bump.yml
@@ -91,7 +91,7 @@ jobs:
         id: changed-files
         uses: tj-actions/changed-files@v36
         with:
-          files: cellxgene_schema_cli/cellxgene_schema/ontology_files/*.csv.gz
+          files: cellxgene_schema_cli/cellxgene_schema/gencode_files/*.csv.gz
       - name: Run Dry-Run Script
         if: steps.changed-files.outputs.any_changed == 'true'
         run: python3 -m scripts.schema_bump_dry_run_genes.gene_bump_dry_run
