+++ cellxgene_schema_cli/cellxgene_schema/gencode_files/.gitignore
@@ -1,0 +1,3 @@
+*owl
+*gtf.gz
+
