--- cellxgene_schema_cli/cellxgene_schema/ontology_files/.gitignore
@@ -1,3 +1,0 @@
-*owl
-*gtf.gz
-
