--- cellxgene_schema_cli/cellxgene_schema/ontology_files/owl_info.yml
@@ -1,39 +1,0 @@
-# Changes to this file will trigger a run of the ontology_process.yml GHA job, which will run the ontology_processing.py script.
-CL:
-  latest: 2024-01-04
-  urls:
-    2024-01-04: https://github.com/obophenotype/cell-ontology/releases/download/v2024-01-04/cl.owl
-EFO:
-  latest: 2024-01-15 EFO 3.62.0
-  urls:
-    2024-01-15 EFO 3.62.0: https://github.com/EBISPOT/efo/releases/download/v3.62.0/efo.owl
-HANCESTRO:
-  latest: 3.0
-  urls:
-    3.0: https://github.com/EBISPOT/hancestro/raw/3.0/hancestro-base.owl
-HsapDv:
-  latest: 2020-03-10
-  urls:
-    2020-03-10: http://aber-owl.net/media/ontologies/HSAPDV/11/hsapdv.owl
-MONDO:
-  latest: 2024-01-03
-  urls:
-    2024-01-03: https://github.com/monarch-initiative/mondo/releases/download/v2024-01-03/mondo.owl
-MmusDv:
-  latest: 2020-03-10
-  urls:
-    2020-03-10: http://aber-owl.net/media/ontologies/MMUSDV/9/mmusdv.owl
-NCBITaxon:
-  latest: 2023-06-20
-  urls:
-    2023-06-20: https://github.com/obophenotype/ncbitaxon/releases/download/v2023-06-20/ncbitaxon.owl.gz
-  children_of:
-    - NCBITaxon:33208
-UBERON:
-  latest: 2024-01-18
-  urls:
-    2024-01-18: https://github.com/obophenotype/uberon/releases/download/v2024-01-18/uberon.owl
-PATO:
-  latest: 2023-05-18
-  urls:
-    2023-05-18: https://github.com/pato-ontology/pato/raw/v2023-05-18/pato.owl
