+++ cellxgene_schema_cli/scripts/readme.md
@@ -3,68 +3,27 @@
 ## Quick start 
 To update the gene and ontology references the following files need to be modified with the updated references:
 
-- Genes `cellxgene_schema_cli/Makfile` (see rules for each organism)
-- Ontologies `cellxgene_schema_cli/cellxgene_schema/ontology_files/owl_info` 
+- `cellxgene_schema_cli/cellxgene_schema/gencode_files/gene_info.yml`
+- Bump cellxgene-ontology-guide version to latest release with updated ontologies in `cellxgene_schema_cli/cellxgene_schema/requirements.txt`
 
-Then the following will command download and process the reference files, and then update the respective package files under `cellxgene_schema_cli/cellxgene_schema/ontology_files/`
-Following an update and release, copy the new ontology files to the single-cell-data-portal repo, replacing the files currently stored in `backend/ontology_files`
+- Then the following will command download and process the reference files, and then update the respective package files under `cellxgene_schema_cli/cellxgene_schema/gencode_files/`
+
 ```bash
 make update-references
 ```
 
-The reference files are used internally by the classes `cellxgene-schema.ontology.GeneChecker` and 
-`cellxgene-schema.ontology.OntologyChecker`, those classes get the references file paths from `cellxgene-schema.env`
-
-For more detailed information please read the rest of this readme.
+The reference files are used internally by the class `cellxgene-schema.ontology.GeneChecker`
 
 ## Updating ontologies
 
-Ontology information for cellxgene-schema is  stored in `cellxgene_schema_cli/cellxgene_schema/ontology_files/all_ontology.json.gz` 
-
-To update these ontologies, a developer has to:
-
-### 1. Modify owl references
-
-The yaml file `cellxgene_schema_cli/cellxgene_schema/ontology_files/owl_info.yml` contains the current ontologies and versions used.
-
-Foe example, for CL:
-
-```yaml
-CL:
-  latest: 2021-06-21
-  urls:
-    2021-06-21: https://github.com/obophenotype/cell-ontology/raw/v2021-06-21/cl.owl
-```
-
-`latest` references the version currently used by cellxgene-schema. When updating an ontology, the new version has to 
-be added to `urls` under the current date and `latest` has to be to point at the correct date. It would like look this:
-
-```yaml
-CL:
-  latest: new_date
-  urls:
-    2021-06-21: https://github.com/obophenotype/cell-ontology/raw/v2021-06-21/cl.owl
-    new_date: https://foo.owl
-```
-
-**Note:** The processing scripts currently support `.owl` and `.owl.gz` files
-
-### 2. Download and parse ontologies
-
-From this directory (`cellxgene_schema_cli/scripts`) the following script has to be called:
-
-```bash
-make download-ontologies
-```
-
-This will:
-
-1. Download ontology owl files to `cellxgene_schema_cli/cellxgene_schema/ontology_files/`
-2. Parse each owl, extract important information and store all of it into **one** file `cellxgene_schema_cli/cellxgene_schema/ontology_files/all_ontology.json`
+Ontology information for cellxgene-schema is derived from cellxgene-ontology-guide. To update the ontologies in 
+conjunction with a schema release, update [ontology_info.json in cellxgene-ontology-guide](https://github.com/chanzuckerberg/cellxgene-ontology-guide/blob/main/ontology-assets/ontology_info.json),
+and then generate a new release with the subsequently generated ontology reference files. Bump the version of cellxgene-ontology-guide in 
+`cellxgene_schema_cli/cellxgene_schema/requirements.txt` to this latest release.
 
 ## Updating gene references
 
-Gene references are stored under  `cellxgene_schema_cli/cellxgene_schema/ontology_files/`
+Gene references are stored under  `cellxgene_schema_cli/cellxgene_schema/gencode_files/`
 
 The following gene references are currently stored (to see specific versions refer to schema definition):
 
@@ -101,5 +60,3 @@ If instead of the rule `update-references` individual make rules were executed, 
 ```bash
 make clean
 ```
-Please ensure you have replaced the ontology files in the single-cell-data-portal repo with  the updated ontology files.
-The files should be stored under `backend/ontology_files`
