--- scripts/compute_mappings/README.md
@@ -1,7 +1,0 @@
-The python scripts in this directory are generated via cli command: 
-
-```
-jupyter nbconvert <notebook_filename>.ipynb --to python
-```
-
-To update scripts, edit the notebooks and re-run the cli command to re-generate the python scripts.
