+++ scripts/schema_bump_dry_run_genes/gene_bump_dry_run.py
@@ -8,7 +8,7 @@ import tiledb as tiledb
 from jinja2 import Template
 
 from cellxgene_schema_cli.cellxgene_schema.env import ONTOLOGY_DIR
-from cellxgene_schema_cli.cellxgene_schema.ontology import SupportedOrganisms
+from cellxgene_schema_cli.cellxgene_schema.gencode import SupportedOrganisms
 from scripts.common.thirdparty.discovery_api import (
     BASE_API,
     fetch_private_collections,
@@ -94,7 +94,7 @@ def get_genes(dataset: dict) -> List[str]:  # type: ignore
 
 
 def get_diff_map() -> Dict[str, List[str]]:
-    # list all of the files ending with diff.txt in the cellxgene_schema/ontology_files directory
+    # list all of the files ending with diff.txt in the cellxgene_schema/gencode_files directory
     # for each file, open it and read the contents into a dictionary
     diff_map = {}
     suffix = "_diff.txt"
